pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "silas"
version = "0.1.0"
description = "Silas runtime skeleton (Phase 1a)"
readme = "PLAN.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8",
  "pydantic-ai-slim[openrouter,logfire]>=0.0.27",
  "pydantic-settings>=2.4",
  "pyyaml>=6.0",
  "click>=8.1",
  "httpx>=0.27",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pynacl>=1.5",
  "cryptography>=43",
  "keyring>=25.2",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.6",
  "aiosqlite>=0.22.1",
  "apscheduler>=3.11,<4",
  "pydantic-ai-backend>=0.1.6",
]

[project.optional-dependencies]
guardrails = ["guardrails-ai"]

[project.scripts]
silas = "silas.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["silas*"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["silas", "tests"]

[tool.ruff.lint]
select = [
  "E", "F", "I", "B", "N", "C901",  # existing
  "UP",   # pyupgrade — modern Python patterns
  "SIM",  # simplify — reduce unnecessary complexity
  "RUF",  # ruff-specific — mutable defaults, ambiguous unicode
  "PT",   # pytest style — fixture/assert best practices
  "S",    # bandit security — hardcoded secrets, unsafe patterns
  "DTZ",  # datetime — enforce timezone-aware datetimes
  "PIE",  # misc — unnecessary pass, duplicate fields
  "T20",  # no print() — use logging instead
]
ignore = [
  "E501",   # line length handled by formatter
  "S101",   # assert in tests is fine
  "S603",   # subprocess calls are intentional in execution layer
  "S607",   # partial executable path — we control the environment
  "UP007",  # X | Y union syntax — keep Optional for readability where needed
]

[tool.ruff.lint.per-file-ignores]
# Tests legitimately use hardcoded values, tmp paths, and broad raises
"tests/*" = ["S104", "S105", "S106", "S108", "PT011", "RUF043", "DTZ001"]
"silas/benchmarks/*" = ["S108"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "benchmark: mark test as a performance benchmark (deselect with '-m \"not benchmark\"')",
]
